Remove parsing dangling empty delegation call
It's attached to some next declaration after whitespace

constructor(a: Any)

<here>fun some() {}

+++ compiler/frontend/src/org/jetbrains/kotlin/psi/KtSecondaryConstructor.kt
@@ -27,27 +27,29 @@ class KtSecondaryConstructor : KtConstructor<KtSecondaryConstructor> {
 
     override fun <R, D> accept(visitor: KtVisitor<R, D>, data: D) = visitor.visitSecondaryConstructor(this, data)
 
-    override fun getContainingClassOrObject() = getParent().getParent() as KtClassOrObject
+    override fun getContainingClassOrObject() = parent.parent as KtClassOrObject
 
     override fun getBodyExpression() = findChildByClass(KtBlockExpression::class.java)
 
     override fun getConstructorKeyword() = notNullChild<PsiElement>(super.getConstructorKeyword())
 
-    fun getDelegationCall(): KtConstructorDelegationCall = findNotNullChildByClass(KtConstructorDelegationCall::class.java)
+    fun getDelegationCall(): KtConstructorDelegationCall? = findChildByClass(KtConstructorDelegationCall::class.java)
 
-    fun hasImplicitDelegationCall(): Boolean = getDelegationCall().isImplicit()
+    fun hasImplicitDelegationCall(): Boolean = getDelegationCall()?.isImplicit ?: true
 
     fun replaceImplicitDelegationCallWithExplicit(isThis: Boolean): KtConstructorDelegationCall {
-        val psiFactory = KtPsiFactory(getProject())
+        val psiFactory = KtPsiFactory(project)
         val current = getDelegationCall()
 
-        assert(current.isImplicit()) { "Method should not be called with explicit delegation call: " + getText() }
-        current.delete()
+        if (current != null) {
+            assert(current.isImplicit) { "Method should not be called with explicit delegation call: " + text }
+            current.delete()
+        }
 
-        val colon = addAfter(psiFactory.createColon(), getValueParameterList())
+        val colon = addAfter(psiFactory.createColon(), valueParameterList)
 
         val delegationName = if (isThis) "this" else "super"
 
-        return addAfter(psiFactory.creareDelegatedSuperTypeEntry(delegationName + "()"), colon) as KtConstructorDelegationCall
+        return addAfter(psiFactory.createDelegatedSuperTypeEntry(delegationName + "()"), colon) as KtConstructorDelegationCall
     }
 }
